Fix Party not updating properly with <6
If the user were to set to the last slot, the setParty method would
trickle the data down, but the getQuickFiller would then make a sprite
appear in the slot it was trickled from.

Since setParty refreshes the party sprites, there's no need to do it for
anything but a party slot.