Add IV/EV click modification
Control to modify IVs, Alt to modify EVs
Left click to set to 31 (or max remaining EVs)
Right click to set to 0.

Feels much better with the Key being the IV/EV switch, not the
left/right click (for left/right column editing).

Think of it as L(oad)/R(eset) :+1: