Add automatic save backup
Create "bak" file, or export a backup successfully anywhere to receive
the "create bak folder" prompt. When saves are loaded (with a new
timestamp), they will be backed up. Repeat loads of the same file will
not be backed up, as a backup of that file already exists
(OT/Game/SavedAt is the hash used).