Re-fix FixMoves
On second thought, that wasn't the best way. Just loop back and fix if
it created a hole. Coul always use a recursive call instead, but I think
this is a little cleaner.

The reverted change didn't fix PP or PP Ups for moves; I didn't think of
this new fix at the time!