Fix nearly identical events with different moves
Charizardite X/Y events in particular came up as invalid (Y).

The Y event had Inferno (not a normal move), but it wasn't in the
relearn pool. Due to these nearly identical events only differing in
terms of usable movepool (not relearn pool), the check has to be able to
be deferred to later if multiple cards are still valid.