Re-add PKM property verification
Item/move/species (didn't add Ball and Version) being imported from a
.pkm or forcefully input now requires the user to OK before it is set to
the save.

EX: Hyperspace Fury isn't in X/Y, drag from another instance of the
program with OR/AS loaded and a Hyperspace Fury move... the user will be
alerted of the incompatibility and will be allowed to continue (and fix,
if desired).