Bug fixes
Box Wallpapers
Stop returning the first box wallpaper regardless of box

EXP:
Changing level/exp wouldn't update stats (EXP needed to be set back to
PKM object)

Drag&Drop:
If the file exists and is locked for editing, the try { } will fail ->
so will delete since the file is locked.
Never seen or heard it happen, but is possible.