Add update available indication
version.txt as a program resource -> check against online copy, if
online has an earlier date then indicate that an update is available.

When posting a new update, the version.txt has to be updated with a new
yyyyMMdd.

Ask me on IRC why I prefer to do things this way ;)