GiveAll medals 0->30->36
Closes #213

Distribution medals only given if the Pokémon has Secret Training unlocked. Clicking Give All with no medals will give 30, then clicking again will give the final 6. Of course if the Pokémon already had it unlocked but not 30 it will bump to 36, but it's up to the user to None/Give as desired.