Hide met tab for gen1 saves, fix verifiedPKM
6->1->translate->6 needs translation done for the Met tab as it being
removed does not list it in the control list when translating, so just
trigger it again.

Other methods had to be tweaked to not reference the tabMain tabs by
index (instead by control names!)

Simplify the verifiedPKM method. removes unnecessary goto and fixes
recent added bug :)
As a side effect the color tracking variables are no longer, now using
the ResetForeColor/ResetBackColor methods of the control -- duh