Remove DVs label, readonly->enable IV
We're gonna call them IVs for gen1/2 because this is {CURRENT YEAR}.
Too much of a hassle to have multiple controls hidden just for
generation specific nomenclature differences.

Change the Readonly for HP & SpD to Enabled (and flip bool states); mask
now appears to enforce non-centering, no need to pad with 0 to fake that
it's centered.

Closes #255, 8 bits down.