Fix uninitialized g4 save loading
Gen4 sets to sav1 (not sav0) when the game is first started, leaving
sav0 uninitialized (0xFF'd).

SaveUtil.cs:
When checking the save, check both sav0 and sav1 for the footer bytes.

SAV4:
When checking for the active blocks, check for this uninitialization
(and return appropriately).
Also fixed the map position offsets (didn't point to sav0/sav1, were
static to sav0).
Minor comment fix (end+=4 to range to match the Take length).