Fix boxviewer / main desync
Communicate between forms to notify when a refresh of the sprites is
needed.

Open up boxviewer to box 1, and have Main on box 1. Move pokemon between
in Main, and the boxviewer now is notified and updates. Only updates if
it is viewing the box that was updated, and wasn't a source/dest that
already initiated its own update.

Even works for dragging pkm files in from another PKHeX.exe or file. No
cross-exe communication.

Realistically, this code should never be triggered, but users may have
multiple copies of the same box viewed (why???).