Fix geolocation checks
Trading Pokemon still sets Geolocation so we expect it as long as an Handling Trainer exists.
Also having a HT but not a location is wrong as long as its not an event.
I only checked the first country slot in last case because its the first one written if the Geolocation list is empty, so having Geo2_Country but not Geo1_Country should be flagged as well. GeoX_Region may be empty depending on the player.