Fix gender symbol setting
if nidoran is not nicknamed, it will use full width if it's not named
Nidoran.
if anything else is nicknamed, it will check to see if any
non-latin/special characters are used to determine if half width should
be set.

possibly applies to gen4/5, if so, please open an issue/post on forums.