Bias rby encounter selection due to lack of info
Previously, if encounterstatic was valid it'd treat it as the actual
encounter.
Now, select the lowest species possible.

Only really affects moon stone evolution families, which are available
in the wild at lower levels than their coin counterpart (so we prefer
them).
Kinda flawed as DV spread can be checked (invalid=>use static) but this
should be good enough for now.