Refactor move legality checking
With the refactoring of result strings outside of Checks.cs, the logic
became harder to follow as the result strings functioned as comments.

This commit splits up the large move checking methods into smaller parts
and added comments.

Moved keldeo move check to a form check and simplifies the check to only
search once.