Fix gen1/2 evo move edge case
Closes #1112
Encounter matched was set to an egg (err 1, core.cs), and the GB
encounter info was not set (err 2, checks.cs).

Fix by preventing eggs if Gen2 is disallowed (reuse allowg1tradeback),
and set the species if it is an egg (future case).

switch pidiv return to null, as it should always be null (not whatever
is returned by the last pidiv finder).