Remove Mirage Island met location from valid gen3
Although Pokémon Ruby, Sapphire, and Emerald include a location pointer
for "Mirage Island", any Wynaut caught here will simply display "Met on
Route 130" in their summary. This is because the player is technically
still on Route 130's map.

Can't get anything with that location ID as an egg or as wild