Fix raid seed editing
Because we pass in a reference to the raid block, rather than obtain the reference from the to-be-edited save file object (which was cloned), we need to restore the unmodified data to the save file instead of copying the modified save data.

Bad explanation, but we've edited the origin save data. Revert to original.