Improve encounterslot yield speed
When we match an evo to the slot, all future evo-species comparisons to that slot will always be false; break instead of continuing.

Remove Pressure Slot level bumping logic for Gen7; each area has a fixed level range shared for all slots, so there'll never be an altform with a higher level.

Sanity check Gen6 pressure slots to ensure that the pressure slot has a different form and the highest level possible for the area.