Refactor: Gen9 Internal->National
In the game data for S/V, the game stores the "Dev ID" and "National Dex ID" in the personal data, and all stored encounter/pkm species IDs use the Dev ID.

In the base ROM (1.0), GameFreak forgot to update the National Dex indexes, and thus the two values were the same (ignoring Spectrier/Glastrier which were swapped). GameFreak eventually fixed this in 1.0.1, but the National Dex was unused and wasn't confirmed until recent merch and media started using the National Dex numbers.

This commit revises the API for DevId<->National conversion, for Gen1 and Gen3 as well. Instead of storing large ushort[] remapping tables, just store a sbyte delta array to shift a range of species values to the correct result. This setup results in no heap allocation on static ctor, as it'll just fetch from the binary's static data. Smaller size stored too :+1:

Adds unit tests for a few a->b and b->a cases.