Don't clear previous flags if form doesn't exist
Closes #3720
All problematic species are form!=0 megas that aren't present in the game
applying seen-all will see the base form (which exists), then un-see the species because form-1 doesn't exist.

just revise the signature & how we check so that it doesn't wipe if we're wanting to set the species, but alter the forms if they were mistakenly set.