Adjust How Glow Is Handled For Gen 9 Sprites
Since there are no shiny artwork sprites, for pre-SV Pokémon the sprite that is shown when shiny is the older one but the glow effect uses the artwork sprite which causes a mismatch. Change the glow effect to use the shiny sprite in gen 9 so that the glow effect matches the Pokémon.
Unfortunate consequence is that the shiny indicator for a sprite has also gets the glow effect.