Add unit test, extract/move logic
If it ever matches \0, it'll return 0 (default), so use that as the return value. Defer the Hiragana check to be done only if the initial remap fails, as checking if it's Hiragana is very unlikely.
Mime is the only Gen1/2 species with a space, so we can quickly check before doing IndexOf; do it only once since the dot is always right before the space we shift down into. Space is never at index 0, so give the compiler a hint that index-1 will always be valid.