Add draft logic for Alph unown, more preference
For each encounter that we may yield, we want to prefer the most likely (no lead) encounter possible otherwise the user will complain "why is everything synchronize and the wrong slot!", so we now retain a priority queue.

Show the encounter conditions & encounter seed on hover. Reworks the PIDIV struct.